+++ NFACT/nfactpp_utils.py
@@ -1,7 +1,5 @@
-import subprocess
 import os
-from nfact_file_checking_functions import colours, make_directory
-import re
+
 
 def add_file_path_for_images(arg: dict, sub: str) -> dict:
     """
@@ -55,31 +53,30 @@ def build_xtract_arguments(arg: dict, sub: str) -> list:
         "-rois",
         rois,
         "-warps",
-        arg["ref"], 
+        arg["ref"],
         images["warps"][0],
         images["warps"][1],
         gpu,
         "-stage",
-        "1", "-tract_list",
+        "1",
+        "-tract_list",
         "null",
-        '-xtract'
+        "-xtract",
     ]
-    
+
     return [Bargs for Bargs in xtract_Bargs if Bargs]
 
-        
-def write_to_file(file_path: str, 
-                  name: str, 
-                  text: str) -> bool:
+
+def write_to_file(file_path: str, name: str, text: str) -> bool:
     """
     Function to write to file.
 
     Parameters
     ----------
     file_path: str
-        abosulte file path to 
-        where file is created 
-    name: str 
+        abosulte file path to
+        where file is created
+    name: str
         name of file
     text: str
         string to add to file
@@ -88,31 +85,43 @@ def write_to_file(file_path: str,
         with open(f"{file_path}/{name}", "w") as file:
             file.write(text)
     except Exception as e:
-        print(f'Unable to write to {file_path}/{name} due to :', e)
+        print(f"Unable to write to {file_path}/{name} due to :", e)
         return False
     return True
 
-def write_options_to_file(file_path: str,  
-                          seed_txt: str):
-    
+
+def write_options_to_file(file_path: str, seed_txt: str):
     """
-    Function to write seeds 
+    Function to write seeds
     and ptx_options to file
 
     Parmeters
     ---------
     file_path: str
-        file path for .PP_config 
-        directory  
+        file path for .PP_config
+        directory
     seed_txt: str
-        path of string to go into 
+        path of string to go into
         seed directory
     """
-    ptx_options = write_to_file(file_path, 
-                                'ptx_options.txt', 
-                                "--pd")
+    ptx_options = write_to_file(file_path, "ptx_options.txt", "--pd")
     if not ptx_options:
         return False
-    seeds = write_to_file(file_path, 'seeds.txt', seed_txt)
+    seeds = write_to_file(file_path, "seeds.txt", seed_txt)
     if not seeds:
         return False
+
+def colours():
+    """
+    Function to print out text in colors
+
+    Parameters
+    ----------
+    None
+
+    Returns
+    -------
+    dict: dictionary object
+        dictionary of color strings
+    """
+    return {"reset": "\033[0;0m", "red": "\033[1;31m"}
